requirements.txt
request
pymongo

colorama
tqdm
re